+++ pkg/qe-tests/cypress/integration/tests/Rhv/config_separate_mapping_rhv.ts
@@ -35,6 +35,7 @@ export const providerData: RhvProviderData = {
   password: v2v_rhv_password,
   cert: v2v_rhv_cert,
 };
+
 // edit rhv provider data
 export const incorrectProviderData: RhvProviderData = {
   type: providerType.rhv,
@@ -44,6 +45,7 @@ export const incorrectProviderData: RhvProviderData = {
   password: 'mtv@123!',
   cert: v2v_rhv_cert,
 };
+
 export const networkMappingPeer: MappingPeer[] = [
   {
     sProvider: 'ovirtmgmt',
@@ -109,7 +111,7 @@ export const editStorageMapping: MappingData = {
 };
 
 export const rhel8Cold: PlanData = {
-  name: 'testplan-rhv-rhel8-separate-mapping-cold',
+  name: `testplan-${providerData.name}separate-mapping-cold`,
   sProvider: providerData.name,
   tProvider: 'host',
   namespace: targetNamespace,
