+++ pkg/qe-tests/cypress/integration/tests/tier0/tier0_config_vmware.ts
@@ -210,171 +210,3 @@ export const vmwareTier0TestArray = [
   vmwareTier0TestCephWarm,
   vmwareTier0TestNfsWarm,
 ];
-
-/**
- * Starting RHV section
- **/
-//Getting required env variables for RHV
-const v2v_rhv_providername = Cypress.env('v2v_rhv_providername');
-const v2v_rhv_username = Cypress.env('v2v_rhv_username');
-const v2v_rhv_password = Cypress.env('v2v_rhv_password');
-const v2v_rhv_hostname = Cypress.env('v2v_rhv_hostname');
-const v2v_rhv_clustername = Cypress.env('v2v_rhv_clustername');
-const v2v_rhv_cert = Cypress.env('v2v_rhv_cert');
-
-//Defining RHV provider
-export const rhvProvider: RhvProviderData = {
-  type: providerType.rhv,
-  name: v2v_rhv_providername,
-  hostname: v2v_rhv_hostname,
-  username: v2v_rhv_username,
-  password: v2v_rhv_password,
-  cert: v2v_rhv_cert,
-};
-
-//Defining RHV network mapping peers for 2 networks
-export const rhvNetworkMappingPeer_2x_network: MappingPeer[] = [
-  {
-    sProvider: 'ovirtmgmt',
-    dProvider: 'Pod network',
-  },
-  {
-    sProvider: 'vm',
-    dProvider: 'default / mybridge',
-  },
-];
-
-//Defining RHV storage mapping peer for NFS
-export const rhvStorageMappingPeer_nfs: MappingPeer[] = [
-  {
-    sProvider: 'v2v-fc',
-    dProvider: storageType.nfs,
-  },
-];
-
-//Defining RHV storage mapping peer for ceph-rbd
-export const rhvStorageMappingPeer_ceph: MappingPeer[] = [
-  {
-    sProvider: 'v2v-fc',
-    dProvider: storageType.cephRbd,
-  },
-];
-
-//Defining RHV network mapping using 2 peers
-export const rhvNetworkMapping_2x_network: MappingData = {
-  name: 'network-qe-rhv-mapping',
-  sProviderName: rhvProvider.name,
-  tProviderName: 'host',
-  mappingPeer: rhvNetworkMappingPeer_2x_network,
-};
-
-//Defining RHV storage mapping for NFS file system
-export const rhvStorageMapping_nfs: MappingData = {
-  name: 'storage-qe-rhv-mapping',
-  sProviderName: rhvProvider.name,
-  tProviderName: 'host',
-  mappingPeer: rhvStorageMappingPeer_nfs,
-};
-
-//Defining RHV storage mapping for ceph-rbd file system
-export const rhvStorageMapping_ceph: MappingData = {
-  name: 'storage-qe-rhv-mapping',
-  sProviderName: rhvProvider.name,
-  tProviderName: 'host',
-  mappingPeer: rhvStorageMappingPeer_ceph,
-};
-
-//Defining RHV cold migration plan for NFS file system
-export const rhvTier0Plan_nfs_cold: PlanData = {
-  name: `rhv-tier0-nfs-${rhvProvider.name}`,
-  sProvider: rhvProvider.name,
-  tProvider: 'host',
-  namespace: 'tier0',
-  sourceClusterName: v2v_rhv_clustername,
-  vmList: vmListArray,
-  useExistingNetworkMapping: true,
-  useExistingStorageMapping: true,
-  providerData: rhvProvider,
-  networkMappingData: rhvNetworkMapping_2x_network,
-  storageMappingData: rhvStorageMapping_nfs,
-  warmMigration: false,
-};
-
-//Defining RHV cold migration plan for ceph-rbd file system
-export const rhvTier0Plan_ceph_cold: PlanData = {
-  name: `rhv-tier0-ceph-${rhvProvider.name}`,
-  sProvider: rhvProvider.name,
-  tProvider: 'host',
-  namespace: 'tier0',
-  sourceClusterName: v2v_rhv_clustername,
-  vmList: vmListArray,
-  useExistingNetworkMapping: true,
-  useExistingStorageMapping: true,
-  providerData: rhvProvider,
-  networkMappingData: rhvNetworkMapping_2x_network,
-  storageMappingData: rhvStorageMapping_ceph,
-  warmMigration: false,
-};
-
-//Defining RHV warm migration plan for NFS file system
-export const rhvTier0Plan_nfs_warm: PlanData = {
-  name: `rhv-tier0-nfs-${rhvProvider.name}`,
-  sProvider: rhvProvider.name,
-  tProvider: 'host',
-  namespace: 'tier0',
-  sourceClusterName: v2v_rhv_clustername,
-  vmList: vmListArray,
-  useExistingNetworkMapping: true,
-  useExistingStorageMapping: true,
-  providerData: rhvProvider,
-  networkMappingData: rhvNetworkMapping_2x_network,
-  storageMappingData: rhvStorageMapping_nfs,
-  warmMigration: true,
-};
-
-//Defining RHV warm migration plan for ceph-rbd file system
-export const rhvTier0Plan_ceph_warm: PlanData = {
-  name: `rhv-tier0-ceph-${rhvProvider.name}`,
-  sProvider: rhvProvider.name,
-  tProvider: 'host',
-  namespace: 'tier0',
-  sourceClusterName: v2v_rhv_clustername,
-  vmList: vmListArray,
-  useExistingNetworkMapping: true,
-  useExistingStorageMapping: true,
-  providerData: rhvProvider,
-  networkMappingData: rhvNetworkMapping_2x_network,
-  storageMappingData: rhvStorageMapping_ceph,
-  warmMigration: true,
-};
-
-//Defining test for RHV cold igration with nfs file system
-export const rhvTier0TestNfsCold: TestData = {
-  loginData: loginData,
-  planData: rhvTier0Plan_nfs_cold,
-};
-
-//Defining test for RHV cold migration with ceph-rbd file system
-export const rhvTiesr0TestCephCold: TestData = {
-  loginData: loginData,
-  planData: rhvTier0Plan_ceph_cold,
-};
-
-//Defining test for RHV wamm igration with nfs file system
-export const rhvTier0TestNfsWarm: TestData = {
-  loginData: loginData,
-  planData: rhvTier0Plan_nfs_warm,
-};
-
-//Defining test for RHV warm migration with ceph-rbd file system
-export const rhvTiesr0TestCephWarm: TestData = {
-  loginData: loginData,
-  planData: rhvTier0Plan_ceph_warm,
-};
-
-export const rhvTier0TestArray = [
-  rhvTier0TestNfsCold,
-  rhvTiesr0TestCephCold,
-  rhvTier0TestNfsWarm,
-  rhvTiesr0TestCephWarm,
-];
