+++ pkg/qe-tests/cypress/integration/tests/tier0/tier0_tests_rhv.ts
@@ -1,20 +1,20 @@
-import { rhvTier0TestArray } from './config_tier0';
-import { login } from '../../../utils/utils';
+import { rhvTier0TestArray } from './tier0_config_rhv';
+import { cleanVms, login } from '../../../utils/utils';
 import { providerRhv } from '../../models/providerRhv';
 import { MappingNetwork } from '../../models/mappingNetwork';
 import { MappingStorage } from '../../models/mappingStorage';
 import { Plan } from '../../models/plan';
 
-describe(
-  'Tier0 tests, creating RHV provider, network and storage(both ceph and nfs) mappings, ' +
-    'plan (both cold and warm), running plan and deleting at the end',
-  () => {
-    const provider = new providerRhv();
-    const networkMapping = new MappingNetwork();
-    const storageMapping = new MappingStorage();
-    const plan = new Plan();
+rhvTier0TestArray.forEach((currentTest) => {
+  describe(
+    'Tier0 tests, creating RHV provider, network and storage(both ceph and nfs) mappings, ' +
+      'plan (both cold and warm), running plan and deleting at the end',
+    () => {
+      const provider = new providerRhv();
+      const networkMapping = new MappingNetwork();
+      const storageMapping = new MappingStorage();
+      const plan = new Plan();
 
-    rhvTier0TestArray.forEach((currentTest) => {
       beforeEach(() => {
         login(currentTest.loginData);
       });
@@ -37,17 +37,12 @@ describe(
       });
 
       after('Deleting plan, mappings and provider created in a previous tests', () => {
-        login(currentTest.loginData);
         plan.delete(currentTest.planData);
         networkMapping.delete(currentTest.planData.networkMappingData);
         storageMapping.delete(currentTest.planData.storageMappingData);
         provider.delete(currentTest.planData.providerData);
-        const namespace = currentTest.planData.namespace;
-        const vm_list = currentTest.planData.vmList;
-        vm_list.forEach((vm) => {
-          cy.exec(`oc delete vm ${vm} -n${namespace}`);
-        });
+        cleanVms(currentTest.planData.vmList, currentTest.planData.namespace);
       });
-    });
-  }
-);
+    }
+  );
+});
